Move MQTT and nats connectors to their own module
