+++ streampipes-extensions/streampipes-connect-adapters-iiot/pom.xml
@@ -41,6 +41,16 @@
         </dependency>
         <dependency>
             <groupId>org.apache.streampipes</groupId>
+            <artifactId>streampipes-connectors-mqtt</artifactId>
+            <version>0.93.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.streampipes</groupId>
+            <artifactId>streampipes-connectors-nats</artifactId>
+            <version>0.93.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.streampipes</groupId>
             <artifactId>streampipes-connectors-pulsar</artifactId>
             <version>0.93.0-SNAPSHOT</version>
         </dependency>
@@ -155,10 +165,6 @@
             <groupId>org.apache.plc4x</groupId>
             <artifactId>plc4j-driver-modbus</artifactId>
             <scope>runtime</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.inlong</groupId>
-            <artifactId>tubemq-client</artifactId>
         </dependency>
         <dependency>
             <groupId>org.bouncycastle</groupId>
