+++ streampipes-extensions/streampipes-connect-adapters/pom.xml
@@ -36,6 +36,11 @@
         <!-- StreamPipes dependencies -->
         <dependency>
             <groupId>org.apache.streampipes</groupId>
+            <artifactId>streampipes-connectors-mqtt</artifactId>
+            <version>0.93.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.streampipes</groupId>
             <artifactId>streampipes-pipeline-elements-shared</artifactId>
             <version>0.93.0-SNAPSHOT</version>
         </dependency>
