+++ streampipes-extensions/streampipes-connectors-nats/src/main/resources/org.apache.streampipes.connect.iiot.protocol.stream.nats/documentation.md
