+++ streampipes-extensions/streampipes-connectors-nats/src/main/resources/org.apache.streampipes.sinks.brokers.jvm.nats/icon.png
